move setupTokenHolders to shared
